I'm not sure how to test this, but I think it'll work. I probably referred to 8 or 9 guides and examples for how to do this before it made sense.
What really helped was learning that I wasn't using tablib syntax correctly. Here's what I coded in a separate file to see if it worked with the Karaoke .mp3/.cdg's we'll be working with.

import taglib

def main():
    song = taglib.File("path")
    print(song.tags['ARTIST'], '-', song.tags['TITLE'])
    print(song)

Printing the tags grabs the metadata, and printing the 'song' prints the full working file path. Knowing this should be helpful for the future.

Let me know if you test this and it doesn't work.